requirements.txt
tensorflow==2.15.0
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
mlflow==2.8.1
python-multipart==0.0.6
onnxruntime==1.16.0